Fix breaking tests, refactor expression tests to their own file
