+++ __TESTS__/unit/actions/Variable.test.ts
@@ -5,49 +5,7 @@ import {Expression} from "../../../src/qualifiers/expression";
 const {set} = Variable;
 const {expression} = Expression;
 
-const cases: Record<string, [string, string]> = {
-  'empty string is not affected': ['', ''],
-  'single space is replaced with a single underscore': [' ', ' '],
-  'blank string is replaced with a single underscore': ['   ', '   '],
-  'underscore is not affected': ['_', '_'],
-  'sequence of underscores and spaces is unchanged': [' _ __  _', ' _ __  _'],
-  'arbitrary text is not affected': ['foobar', 'foobar'],
-  'duration is recognized as a variable and replaced with du': ['duration', 'du'],
-  'double ampersand replaced with and operator': ['foo && bar', 'foo_and_bar'],
-  'double ampersand with no space at the end is not affected': ['foo&&bar', 'foo&&bar'],
-  'width recognized as variable and replaced with w': ['width', 'w'],
-  'initial aspect ratio recognized as variable and replaced with iar': ['initial_aspect_ratio', 'iar'],
-  '$width recognized as user variable and not affected': ['$width', '$width'],
-  '$initial_aspect_ratio recognized as user variable followed by aspect_ratio variable': [
-    '$initial_aspect_ratio',
-    '$initial_ar'
-  ],
-  '$mywidth recognized as user variable and not affected': ['$mywidth', '$mywidth'],
-  '$widthwidth recognized as user variable and not affected': ['$widthwidth', '$widthwidth'],
-  '$_width recognized as user variable and not affected': ['$_width', '$_width'],
-  '$__width recognized as user variable and not affected': ['$__width', '$__width'],
-  '$$width recognized as user variable and not affected': ['$$width', '$$width'],
-  '$height recognized as user variable and not affected': ['$height_100', '$height_100'],
-  '$heightt_100 recognized as user variable and not affected': ['$heightt_100', '$heightt_100'],
-  '$$height_100 recognized as user variable and not affected': ['$$height_100', '$$height_100'],
-  '$heightmy_100 recognized as user variable and not affected': ['$heightmy_100', '$heightmy_100'],
-  '$myheight_100 recognized as user variable and not affected': ['$myheight_100', '$myheight_100'],
-  '$heightheight_100 recognized as user variable and not affected': [
-    '$heightheight_100',
-    '$heightheight_100'
-  ],
-  '$theheight_100 recognized as user variable and not affected': ['$theheight_100', '$theheight_100'],
-  '$__height_100 recognized as user variable and not affected': ['$__height_100', '$__height_100']
-};
-
 describe('Tests for Transformation Action -- Variable', () => {
-  it('tests expressions values', () => {
-    Object.keys(cases).forEach(function (testDescription) {
-      const [input, expected] = cases[testDescription];
-      expect(expression(input).toString()).toBe(expected);
-    });
-  });
-
   it('tests common variable values', () => {
     expect(set('a', 30).toString()).toBe('$a_30');
     expect(set('a', '30').toString()).toBe('$a_!30!');
